This spinet of code that activates the PCs webcam and collects 10 images storing them in a folder
Opens the webcam, creates a database: a folder containing the images of the person captured.The folder is named by the user prompt. The input captured is saved as the data file name "example". the name of the folder is the one used to identify the person while the persons sample images are numbered "exampled001", "example002" and so on containing an array of images to be used by the model to train for facial recognition